+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
  * `Position` is no longer a typedef of `geo_types::Point`. There is also a new, more explicit
    way of constructing it - `from_lat_lon`, and `from_lon_lat`.
+ * If center position is detached, zooming using mouse wheel will now keep location under pointer
+   fixed.
 
 ## 0.12.0